[YSA-81] Feature: ui 수정
