+++ gradle/libs.versions.toml
@@ -18,12 +18,10 @@ exifinterfaceVersion = "1.3.7"
 firebaseMessagingKtx = "24.1.2"
 glanceAppwidget = "1.1.1"
 glanceMaterial3 = "1.1.1"
-kotlinxCoroutinesPlayServices = "1.7.3"
 pagingCompose = "3.3.4"
 
 ## Kotlin
 # https://github.com/JetBrains/kotlin
-coilNetworkOkhttp = "3.2.0"
 kotlin = "2.2.0"
 # https://github.com/Kotlin/kotlinx.serialization
 kotlinxSerializationJson = "1.9.0"
